Exception Visibility
I've added a Moltin\SDK\Exception\Exception interface in the library as
this allows clients to catch any Moltin specific Exception raised by
the library.

In addition to this, I've changed the Exceptions to extend \RunTimeException
as the Exceptions can only be found during runtime.

There should be no breaking changes.